Enable '-Xfatal-warnings' compiler flag (#525)
* Enable '-Xfatal-warnings' flag

* Removed method type parameters

* More test warning fixes 

* Small one

* Address integration test warnings

* Small typo

* Removed cats effect from dependencies

* Add mima exclusion

+++ gcs/src/main/scala/monix/connect/gcp/storage/configuration/GcsBlobInfo.scala
@@ -71,7 +71,7 @@ object GcsBlobInfo {
         List.empty[Acl]
       },
       owner = Option(blobInfo.getOwner),
-      size = Option(blobInfo.getSize()),
+      size = Option(blobInfo.getSize.toLong),
       contentType = Option(blobInfo.getContentType),
       contentEncoding = Option(blobInfo.getContentEncoding),
       contentDisposition = Option(blobInfo.getContentDisposition),
